+++ packages/ocean-react-formik/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.71.0](https://github.com/ocean-ds/ocean-web/compare/v1.70.1...v1.71.0) (2024-12-06)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
+## [1.70.1](https://github.com/ocean-ds/ocean-web/compare/v1.70.0...v1.70.1) (2024-12-04)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
+# [1.70.0](https://github.com/ocean-ds/ocean-web/compare/v1.69.1...v1.70.0) (2024-12-04)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
 ## [1.69.1](https://github.com/ocean-ds/ocean-web/compare/v1.69.0...v1.69.1) (2024-11-13)
 
 **Note:** Version bump only for package @useblu/ocean-react-formik